+++ app/routes/components/ScrollArea/changelog.codegen.ts
@@ -6,4 +6,6 @@ export const changelog = ''
   + '- 2025-10-05: Make the scrollbar appear on hover only.\n'
   + '- 2025-10-10: Add a dragging state for the scrollbar thumb to prevent it from\n'
   + '  hiding while being dragged.\n'
+  + '- 2025-10-14: Update the style of the scrollbar thumb - make sure it has more\n'
+  + '  alpha when dragging.\n'
   ;