+++ src/components/ScrollArea/index.ts
@@ -1,3 +1,3 @@
-// ScrollArea from copy-ui @ 2025-10-10.
+// ScrollArea from copy-ui @ 2025-10-14.
 
 export { ScrollArea } from "./ScrollArea";