+++ src/components/ScrollArea/CHANGELOG.md
@@ -2,3 +2,5 @@
 - 2025-10-05: Make the scrollbar appear on hover only.
 - 2025-10-10: Add a dragging state for the scrollbar thumb to prevent it from
   hiding while being dragged.
+- 2025-10-14: Update the style of the scrollbar thumb - make sure it has more
+  alpha when dragging.